feat(client): restructure graphic module to use SFML submodule and improve organization
