+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "src/client/Graphic/SFML"]
 	path = src/client/Graphic/SFML
 	url = git@github.com:SFML/SFML.git
+[submodule "src/client/GraphicSFML/SFML"]
+	path = src/client/GraphicSFML/SFML
+	url = git@github.com:SFML/SFML.git
